Modified the navbar even more, outsourced some components
