+++ src/components/SearchSuggestionItem.vue
@@ -1,0 +1,11 @@
+<script setup>
+</script>
+
+<template>
+    <div>
+
+    </div>
+</template>
+
+<style scoped>
+</style>
